DFlash2: run the draft model in bfloat16 under a float16 target
Opt-in via DF2_DRAFT_BF16=1. Without it, DFlash2 acceptance on this stack is
0.08% -- every draft logit is NaN -- while bfloat16 gives ~45%.

ROOT CAUSE, measured, not inferred. The drafter's own residual stream inside its
5 decoder layers reaches 1.6e5 and its layer outputs 8.7e4, both past float16's
65504 ceiling. Probed per boundary under fp16:

  fc.out              1.32e4  clean       hidden_norm out   6.379  clean
  layer0.residual     2.71e4  clean
  layer1.residual     5.40e4  inf x8   <- born here
  layer1.hidden          NaN          -> every later layer NaN

So it is exponent RANGE, not precision, and not accumulation. Ruled out by
measurement: the drafter's WEIGHTS are entirely fp16-safe (max|x| 20.9 over all
83 tensors, zero overflow); fp32 accumulation in _score_edges alone does not
help; and carrying only the residual in fp32 merely moves the overflow to the
layer output (0.04% -> 0.88%, still broken). Note this is a DIFFERENT SITE from
the published reports (z-lab/omlx-fork#2, ruixiang63/llama.cpp#2), which describe
an fc GEMM overflow feeding RMSNorm -- that mechanism is real but is not this one.

Four dtype boundaries need casts, because the drafter shares tensors with the
fp16 target. All are no-ops when the dtypes already match:
  - dflash/utils.py: make the DRAFT's model_config current during construction.
    get_model() passes draft_model_config to the LOADER, but the model's __init__
    reads get_current_vllm_config().model_config, which is still the TARGET's --
    so fc, mask_embedding and the candidate selector would be built at the
    target's dtype and tear against bf16 weights. Patching only the dtype in
    SpeculativeConfig is NOT sufficient for this reason.
  - qwen3_dflash.py combine_hidden_states: aux hidden states arrive from the
    target. This is the guard vLLM #40334 proposed and upstream closed without
    merging; ROCm hits it even in the same-dtype case (#42588).
  - qwen3_dflash.py forward: the drafter SHARES the target's embedding table.
  - qwen3_dflash.py _project_context_kv: the speculator's hidden_states buffer is
    allocated at the target's dtype.
  - qwen3_dflash2.py compute_candidates: runs draft hidden states through the
    TARGET's lm_head.

Also requires --kv-cache-dtype float16 at runtime. kv_cache_interface.py asserts
that every layer in a KV cache group is identical, dtype included, and draft
layers land in the same unified spec dict; with cache_dtype "auto" a bf16 draft
attention layer aborts at grouping. An explicit dtype string is applied to every
layer regardless of each model_config, which satisfies the assert at no cost --
the target is fp16 anyway.

MEASURED on 4x RX 7900 XTX, Qwen3.8-27B INT8 W8A16 target + Exllama, TP=4:
acceptance 45.24% (fp16 control: 0.00%), prefix reuse intact (t2/t1 = 0.19).
NOT deployed here: at depth on prose MTP still wins (-13/-9/-3% at 8k/29k/180k);
DFlash2 wins code at depth 0 (126.1 vs 105.5 t/s). Code-at-depth is unmeasured.

+++ vllm/model_executor/models/qwen3_dflash2.py
@@ -282,8 +282,13 @@ class DFlash2Qwen3ForCausalLM(DFlashQwen3ForCausalLM):
     def compute_candidates(
         self, hidden_states: torch.Tensor
     ) -> tuple[torch.Tensor, torch.Tensor]:
+        # This runs the DRAFTER's hidden states through the TARGET's lm_head, so
+        # a bf16 draft under an fp16 target meets an fp16 head here. Post-norm
+        # values are O(10), so the cast is lossless; no-op when dtypes match.
         return self.candidate_logits_processor.get_top_k_tokens(
-            self.lm_head, hidden_states, self.model.candidate_selector.top_k
+            self.lm_head,
+            hidden_states.to(self.lm_head.weight.dtype),
+            self.model.candidate_selector.top_k,
         )
